# Dominator analysis: analyse on basic block granularity