--- a/src/analyses/natural_loops.h
+++ b/src/analyses/natural_loops.h
@@ -31,22 +31,30 @@ Author: Georg Weissenbacher, georg@weissenbacher.name
 template <class, class>
 class natural_loops_templatet;
 
-/// A natural loop, specified as a set of instructions
+/// A natural loop, specified as a set of basic block indices that correspond
+/// to graph nodes in the linked \ref natural_loop_templatet's dominator
+/// analysis.
 template <class P, class T>
 class natural_loop_templatet
 {
   typedef natural_loops_templatet<P, T> natural_loopst;
-  // For natural_loopst to directly manipulate loop_instructions, cf. clients
-  // which should use the public iterface:
+  // For natural_loopst to directly manipulate loop_blocks, cf. clients
+  // which should use the public interface:
   friend natural_loopst;
 
-  typedef std::set<T> loop_instructionst;
-  loop_instructionst loop_instructions;
+  typedef std::set<std::size_t> loop_blockst;
+  loop_blockst loop_blocks;
 
 public:
   explicit natural_loop_templatet(natural_loopst &natural_loops)
     : natural_loops(natural_loops)
   {
+  }
+
+  /// Returns true if \p block_index is in this loop
+  bool contains(std::size_t block_index) const
+  {
+    return loop_blocks.count(block_index);
   }
 
   /// Returns true if \p instruction is in this loop
@@ -67,38 +75,40 @@ public:
   }
 
   // NOLINTNEXTLINE(readability/identifiers)
-  typedef typename loop_instructionst::const_iterator const_iterator;
+  typedef typename loop_blockst::const_iterator const_iterator;
 
-  /// Iterator over this loop's instructions
+  /// Iterator over this loop's blocks
   const_iterator begin() const
   {
-    return loop_instructions.begin();
+    return loop_blocks.begin();
   }
 
-  /// Iterator over this loop's instructions
+  /// Iterator over this loop's blocks
   const_iterator end() const
   {
-    return loop_instructions.end();
+    return loop_blocks.end();
   }
 
-  /// Number of instructions in this loop
+  /// Number of blocks in this loop
   std::size_t size() const
   {
-    return loop_instructions.size();
+    return loop_blocks.size();
   }
 
-  /// Returns true if this loop contains no instructions
+  /// Returns true if this loop contains no blocks
   bool empty() const
   {
-    return loop_instructions.empty();
+    return loop_blocks.empty();
   }
 
-  /// Adds \p instruction to this loop. The caller must verify that the added
-  /// instruction does not alter loop structure; if it does they must discard
-  /// and recompute the related \ref natural_loopst instance.
-  void insert_instruction(const T instruction)
+  const loop_blockst &blocks() const
   {
-    loop_instructions.insert(instruction);
+    return loop_blocks;
+  }
+
+  const std::vector<T> &get_basic_block(std::size_t index) const
+  {
+    return natural_loops.get_basic_block(index);
   }
 
 private:
@@ -141,16 +151,63 @@ public:
     return cfg_dominators;
   }
 
-  /// Returns true if \p instruction is in \p loop
-  bool loop_contains(const natural_loopt &loop, const T instruction) const
-  {
-    return loop.loop_instructions.count(instruction);
-  }
-
   /// Returns true if \p instruction is the header of any loop
   bool is_loop_header(const T instruction) const
   {
     return loop_map.count(instruction);
+  }
+
+  /// Returns true if \p block_index is the header block of any loop
+  bool is_loop_header_block(std::size_t block_index) const
+  {
+    return is_loop_header(get_basic_block(block_index).front());
+  }
+
+  /// Returns the basic block with the given \p index
+  const std::vector<T> &get_basic_block(std::size_t index) const
+  {
+    return cfg_dominators.cfg[index].block;
+  }
+
+  /// Returns true if \p instruction is in \p loop
+  bool loop_contains(const natural_loopt &loop, const T instruction) const
+  {
+    const auto instruction_block = cfg_dominators.get_node_index(instruction);
+    return loop.loop_blocks.count(instruction_block);
+  }
+
+  /// Inserts instruction \p new_instruction after \p insert_after. The new
+  /// instruction must join the same basic block as \p insert_after (i.e.
+  /// \p insert_after must not be a block terminator), such that the insertion
+  /// has no effect on loop structure.
+  void insert_instruction_after(const T new_instruction, const T insert_after)
+  {
+    cfg_dominators.cfg.insert_instruction_after(new_instruction, insert_after);
+  }
+
+  /// Splits the basic block containing \p instruction immediately after
+  /// \p instruction. The caller should fix up the new block; the new block will
+  /// be added to any loops that contained the old one.
+  void split_basic_block_after(const T instruction)
+  {
+    auto new_block_index =
+      cfg_dominators.cfg.split_basic_block_after(instruction);
+    auto old_block_index = cfg_dominators.cfg.entry_map.at(instruction);
+    for(auto &loop : loop_map)
+    {
+      if(loop.second.loop_blocks.count(old_block_index))
+        loop.second.loop_blocks.insert(new_block_index);
+    }
+  }
+
+  /// Adds a new basic block graph edge, which must not alter loop structure,
+  /// as we make no attempt to maintain analysis consistency here.
+  void
+  add_basic_block_graph_edge(const T from_block_tail, const T to_block_head)
+  {
+    // We assume for now that these don't alter loop structure.
+    cfg_dominators.cfg.add_basic_block_graph_edge(
+      from_block_tail, to_block_head);
   }
 
   natural_loops_templatet()
@@ -238,30 +295,31 @@ void natural_loops_templatet<P, T>::compute_natural_loop(T m, T n)
 {
   assert(n->location_number<=m->location_number);
 
-  std::stack<T> stack;
+  std::stack<std::size_t> stack;
 
   auto insert_result = loop_map.emplace(n, natural_loopt{*this});
   INVARIANT(insert_result.second, "each loop head should only be visited once");
   natural_loopt &loop = insert_result.first->second;
 
-  loop.insert_instruction(n);
-  loop.insert_instruction(m);
+  const auto n_block = cfg_dominators.cfg.entry_map.at(n);
+  const auto m_block = cfg_dominators.cfg.entry_map.at(m);
+  loop.loop_blocks.insert(n_block);
+  loop.loop_blocks.insert(m_block);
 
-  if(n!=m)
-    stack.push(m);
+  if(n_block != m_block)
+    stack.push(m_block);
 
   while(!stack.empty())
   {
-    T p=stack.top();
+    std::size_t p = stack.top();
     stack.pop();
 
-    const nodet &node = cfg_dominators.get_node(p);
+    const nodet &node = cfg_dominators.cfg[p];
 
     for(const auto &edge : node.in)
     {
-      T q=cfg_dominators.cfg[edge.first].PC;
-      std::pair<typename natural_loopt::const_iterator, bool> result =
-        loop.loop_instructions.insert(q);
+      std::size_t q = edge.first;
+      auto result = loop.loop_blocks.insert(q);
       if(result.second)
         stack.push(q);
     }
@@ -277,12 +335,17 @@ void natural_loops_templatet<P, T>::output(std::ostream &out) const
     unsigned n=loop.first->location_number;
 
     out << n << " is head of { ";
-    for(typename natural_loopt::const_iterator l_it=loop.second.begin();
-        l_it!=loop.second.end(); ++l_it)
+    bool first = true;
+    for(const auto loop_block : loop.second.loop_blocks)
     {
-      if(l_it!=loop.second.begin())
-        out << ", ";
-      out << (*l_it)->location_number;
+      for(const auto &instruction : this->get_basic_block(loop_block))
+      {
+        if(!first)
+          out << ", ";
+        else
+          first = false;
+        out << instruction->location_number;
+      }
     }
     out << " }\n";
   }