… src/goto-instrument/goto_instrument_parse_options.cpp
… src/goto-instrument/goto_program2code.cpp
… src/goto-instrument/loop_utils.cpp
… src/goto-instrument/reachability_slicer.cpp
… src/goto-programs/cfg.h
… src/goto-programs/goto_program.h
… src/util/dense_integer_map.h