@smowton 2019-08-21
@hannes-steffenhagen-diffblue take another look, I had to make the get-basic-block-from-instruction operation cheaper to get acceptable performance, for which I introduced `dense_integer_mapt`.

Preliminary performance figures, using the same benchmarking example as you did in https://github.com/diffblue/cbmc/pull/1787#issuecomment-362337130

Examples:
* `straightline` is like `v1` in the linked comment: a long straight-line function.
* `small`, `medium` and `largeish` are 100, 1000 and 3000 repeats of example `v2` in the linked comment (a function with lots of small basic blocks)
* `blocks` is a version of `medium` with basic blocks of size 3

As with the linked comment, I used `dependency-graph` to benchmark as a pass that both computes postdoms and then hammers it with queries (58% of runtime is spent in `control_dependencies` of which almost all of that is postdom queries).

Results:

* Straightline: 4.5s -> 0.5s (down 89%)
* Blocks: 7.5s -> 6.0s (down 20%)
* Small: 0.05s -> 0.04s (unmeasurable)
* Medium: 4.6s -> 3.3s (down 28%)
* Largeish: 107s -> 61s (down 43%)

So up front this looks to me like we should get a huge win for mostly-straight-line code (as you'd expect from using basic-block form), and moderate wins (20-30%) from code consisting of 2-3 line basic blocks. Unlike #1787 we continue to use Cooper's dominator algorithm, but we shrink the sizes of the sets it tracks using basic-block form, which particularly shows its worth when we deal with large GOTO programs.