@smowton 2019-08-21
re: breaking up the PR, a lot of this now introduces and uses a sensible API for dominators, cf. the original code which merrily rummaged in both `cfg_dominatorst` and `cfgt`'s fields. If this is looking like it might go in then I'll submit those first as cleanups.