@smowton 2019-08-21
Best case against the Java testsuite: `regression/jbmc/coreModels`, down from 3.2s to 1.7s. Total `jbmc-CORE` test suite time down from 67 to 59 seconds.