@smowton 2019-08-22
Have so far split out #5045 #5051 #5055, which should be pretty uncontroversial, and #5049, which floats the `dense_integer_mapt` concept for `cfgt` by itself, without the basic-blocks innovation introduced here. I will make one more PR in the morning, to factor out this PR's cleanup changes to `natural_loopst`, then this one will sit on top of the whole stack.